refactor: split RedisZoneRegistry into Classic/Instanced strategy classes (#370)
Replace the single RedisZoneRegistry class (which branched on an
`instancing` flag in every public method) with two focused classes:

- ClassicRedisZoneRegistry: simple key-value zone ownership
- InstancedRedisZoneRegistry: hash-based multi-engine zone instances

The composition roots (MudServer, EngineServer, GatewayServer) now
select the appropriate class at construction time based on config.
