+++ src/main/kotlin/dev/ambon/sharding/InstancedRedisZoneRegistry.kt
@@ -8,70 +8,27 @@ import io.github.oshai.kotlinlogging.KotlinLogging
 private val log = KotlinLogging.logger {}
 
 /**
- * Zone registry backed by Redis keys with TTL. Each engine writes its zone
- * claims as Redis keys and refreshes them periodically. If an engine dies,
- * its leases expire after [leaseTtlSeconds].
+ * Zone registry backed by Redis hashes with per-engine lease keys, instanced mode.
+ * Multiple engines may claim the same zone; each (zone, engineId) pair is a
+ * distinct [ZoneInstance] with its own player count and capacity.
  *
- * When [instancing] is `true`, multiple engines may claim the same zone.
- * Instance data (including player counts) is stored in Redis hashes keyed
- * by zone, with one field per engine.
- *
- * Redis key scheme (classic, non-instanced):
- * - `zone:owner:<zone>` → JSON [EngineAddress] (TTL = leaseTtlSeconds)
- *
- * Redis key scheme (instanced):
+ * Redis key scheme:
  * - `zone:instances:<zone>` → HASH { engineId → JSON [ZoneInstance] }
  * - `zone:lease:<zone>:<engineId>` → "1" (TTL = leaseTtlSeconds)
  */
-class RedisZoneRegistry(
+class InstancedRedisZoneRegistry(
     private val redis: RedisConnectionManager,
     private val mapper: ObjectMapper,
     private val leaseTtlSeconds: Long = 30L,
-    private val keyPrefix: String = "zone:owner:",
-    private val instancing: Boolean = false,
     private val defaultCapacity: Int = ZoneInstance.DEFAULT_CAPACITY,
 ) : ZoneRegistry {
     private val instanceHashPrefix = "zone:instances:"
     private val leasePrefix = "zone:lease:"
 
-    override fun ownerOf(zone: String): EngineAddress? {
-        if (instancing) {
-            return instancesOf(zone).firstOrNull()?.address
-        }
-        val json = redis.get("$keyPrefix$zone") ?: return null
-        return try {
-            mapper.readValue<EngineAddress>(json)
-        } catch (e: Exception) {
-            log.warn(e) { "Failed to deserialize zone owner for '$zone'" }
-            null
-        }
-    }
+    override fun ownerOf(zone: String): EngineAddress? =
+        instancesOf(zone).firstOrNull()?.address
 
     override fun claimZones(
-        engineId: String,
-        address: EngineAddress,
-        zones: Set<String>,
-    ) {
-        if (instancing) {
-            claimZonesInstanced(engineId, address, zones)
-        } else {
-            claimZonesClassic(engineId, address, zones)
-        }
-    }
-
-    private fun claimZonesClassic(
-        engineId: String,
-        address: EngineAddress,
-        zones: Set<String>,
-    ) {
-        val json = mapper.writeValueAsString(address)
-        for (zone in zones) {
-            redis.setEx("$keyPrefix$zone", leaseTtlSeconds, json)
-        }
-        log.info { "Engine '$engineId' claimed zones: $zones (TTL=${leaseTtlSeconds}s)" }
-    }
-
-    private fun claimZonesInstanced(
         engineId: String,
         address: EngineAddress,
         zones: Set<String>,
@@ -94,41 +51,6 @@ class RedisZoneRegistry(
 
     override fun renewLease(engineId: String) {
         val commands = redis.commands ?: return
-        if (instancing) {
-            renewLeaseInstanced(engineId, commands)
-        } else {
-            renewLeaseClassic(engineId, commands)
-        }
-    }
-
-    private fun renewLeaseClassic(
-        engineId: String,
-        commands: io.lettuce.core.api.sync.RedisCommands<String, String>,
-    ) {
-        val cursor =
-            commands.scan(
-                io.lettuce.core.ScanArgs.Builder
-                    .matches("$keyPrefix*")
-                    .limit(100),
-            )
-        for (key in cursor.keys) {
-            val json = commands.get(key) ?: continue
-            val addr =
-                try {
-                    mapper.readValue<EngineAddress>(json)
-                } catch (_: Exception) {
-                    continue
-                }
-            if (addr.engineId == engineId) {
-                commands.expire(key, leaseTtlSeconds)
-            }
-        }
-    }
-
-    private fun renewLeaseInstanced(
-        engineId: String,
-        commands: io.lettuce.core.api.sync.RedisCommands<String, String>,
-    ) {
         val cursor =
             commands.scan(
                 io.lettuce.core.ScanArgs.Builder
@@ -141,36 +63,6 @@ class RedisZoneRegistry(
     }
 
     override fun allAssignments(): Map<String, EngineAddress> {
-        if (instancing) {
-            return allAssignmentsInstanced()
-        }
-        return allAssignmentsClassic()
-    }
-
-    private fun allAssignmentsClassic(): Map<String, EngineAddress> {
-        val commands = redis.commands ?: return emptyMap()
-        val result = mutableMapOf<String, EngineAddress>()
-        val cursor =
-            commands.scan(
-                io.lettuce.core.ScanArgs.Builder
-                    .matches("$keyPrefix*")
-                    .limit(1000),
-            )
-        for (key in cursor.keys) {
-            val zone = key.removePrefix(keyPrefix)
-            val json = commands.get(key) ?: continue
-            val addr =
-                try {
-                    mapper.readValue<EngineAddress>(json)
-                } catch (_: Exception) {
-                    continue
-                }
-            result[zone] = addr
-        }
-        return result
-    }
-
-    private fun allAssignmentsInstanced(): Map<String, EngineAddress> {
         val commands = redis.commands ?: return emptyMap()
         val result = mutableMapOf<String, EngineAddress>()
         val cursor =
@@ -183,7 +75,6 @@ class RedisZoneRegistry(
             val zone = key.removePrefix(instanceHashPrefix)
             val entries = commands.hgetall(key) ?: continue
             for ((engineId, json) in entries.toSortedMap()) {
-                // Only include if lease is still alive
                 val leaseKey = "$leasePrefix$zone:$engineId"
                 if (commands.exists(leaseKey) > 0) {
                     val instance =
@@ -192,7 +83,6 @@ class RedisZoneRegistry(
                         } catch (_: Exception) {
                             continue
                         }
-                    // First live instance wins for allAssignments (1:1 compat)
                     result.putIfAbsent(zone, instance.address)
                 }
             }
@@ -201,13 +91,6 @@ class RedisZoneRegistry(
     }
 
     override fun instancesOf(zone: String): List<ZoneInstance> {
-        if (!instancing) {
-            return listOfNotNull(
-                ownerOf(zone)?.let {
-                    ZoneInstance(engineId = it.engineId, address = it, zone = zone)
-                },
-            )
-        }
         val commands = redis.commands ?: return emptyList()
         val entries = commands.hgetall("$instanceHashPrefix$zone") ?: return emptyList()
         return entries
@@ -226,7 +109,6 @@ class RedisZoneRegistry(
         engineId: String,
         zoneCounts: Map<String, Int>,
     ) {
-        if (!instancing) return
         val commands = redis.commands ?: return
         for ((zone, count) in zoneCounts) {
             val hashKey = "$instanceHashPrefix$zone"
@@ -242,5 +124,5 @@ class RedisZoneRegistry(
         }
     }
 
-    override fun instancingEnabled(): Boolean = instancing
+    override fun instancingEnabled(): Boolean = true
 }
